refactored code into modular parts #40
