+++ flaskr/auth.py
@@ -2,7 +2,7 @@ import functools
 from flask import Blueprint
 from flask import url_for, redirect, render_template
 from flask import request, session
-from user import User
+from flaskr.user import User
 
 
 bp = Blueprint("auth", __name__, url_prefix="/auth")
@@ -22,7 +22,7 @@ def login_required(view):
 def login():
     error = None
     if "user" in session:
-        return redirect(url_for("inbox"))
+        return redirect(url_for("home"))
 
     if request.method == "POST":
         email = request.form["email"]
@@ -32,7 +32,7 @@ def login():
         error = user_obj.validate()
         if error is None:
             session["user"] = [email, password]
-            return redirect(url_for("inbox"))
+            return redirect(url_for("home"))
 
     return render_template("login.html", error=error)
