--- requirements.txt
@@ -1,22 +1,0 @@
-atomicwrites==1.4.0
-attrs==21.2.0
-click==8.0.1
-colorama==0.4.4
-dnspython==2.1.0
-email-validator==1.1.3
-Flask==2.0.1
-idna==3.3
-iniconfig==1.1.1
-itsdangerous==2.0.1
-Jinja2==3.0.1
-mail-parser==3.15.0
-MarkupSafe==2.0.1
-packaging==21.0
-pluggy==1.0.0
-py==1.10.0
-pyparsing==2.4.7
-pytest==6.2.5
-simplejson==3.17.5
-six==1.16.0
-toml==0.10.2
-Werkzeug==2.0.1
