+++ setup.py
@@ -1,0 +1,12 @@
+from setuptools import setup, find_packages
+
+setup(
+    name="flaskr",
+    version="1.0.0",
+    packages=find_packages(),
+    install_requires=[
+        "flask",
+        "email-validator",
+        "mail-parser",
+    ],
+)
