--- templates/forward.html
@@ -1,38 +1,0 @@
-{% extends "base.html" %}
-{% block title %}Forward Page{% endblock %}
-{% block content %}
-<!-- TODO: TODO: Add nice layout for the composition page -->
-<head>
-  <link rel="stylesheet" href="{{ url_for('static', filename='CSS/composition.css') }}">
-</head>
-
-  <h2> Forward an email </h2>
-
-<body>
-<p><a href="/inbox">Go back to inbox</a></p>
-<p><a href="{{ url_for('inbox_message', uid=uid) }}">Go back to message</a></p>
-<form action="#" method="post" enctype=multipart/form-data>
-  <!-- Using the input and text area to get the user's input via the browser -->
-  <p><input type="text" placeholder="To: " name="receiver" required/></p>
-  <p><input type="text" placeholder="Subject: " name="subject" value="Fwd: {{ subject }}"/></p>
-
-  <!-- Using the Tiny MCE UI to improve the email composition -->
-  <script
-          src="https://cdn.tiny.cloud/1/vr2x3cxjja3hokctzkzvlvxa1ex2uerrs733r5889trcniq3/tinymce/5/tinymce.min.js" referrerpolicy="origin">
-  </script>
-
-  <script>
- tinymce.init({
-  selector: 'textarea',  // change this value according to your HTML
-  menubar: 'edit view format'
-});
-</script>
-
-  <p><textarea name=emailbody rows="40" cols="200">{{ text }}</textarea></p>
-  <p><label for="infile">Attach a file:</label> </p>
-  <input type="file" id="infile" name="infile" multiple=""><br>
-
-  <p><input type="submit" id="Send message" name="Send message"/></p>
-</form>
-</body>
-{% endblock %}
