+++ tests/test_example.py
@@ -2,3 +2,4 @@
 def test_example():
     """This function exists to pass the test, but it should be removed when we implement the tests """
     assert 1 == 1
+
